Disable gold in mozconfig
After switching to profile amd64/23.0 I've discovered that palemoon failed
to rebuild. It turned out that Palemoon uses ld.gold for linking by default
if it's found in the system (which is true in my case), and new 23.0
profile now has -Wl,-z,pack-relative-relocs in LDFLAGS, which is not
supported by ld.gold. It seems to me that disabling gold would be the
right thing to do, as it can only lead to slightly larger linking times
which is not crucial here.